refactor(trpc): file-organization sweep over the loose routers, guards, procedures, plugins and middleware
The two context input shapes are model files; Context itself stays beside createContext as the type derived from it. AuthedProcedureMap moves under procedure/room, the one subfolder that reads it.
